Add a random number generator to sensor pt 1
 Summary:
Some candidates struggle with making API calls. It's difficult to tell whether this is a stack familiarity issue or not, so let's move that portion to the end. Since I think we get more valuable signal from the portion afterwards (i.e. decisions in how to show the data in the UI, how to limit to ten values, how rotations are handled), this helps us get that signal earlier.

Note: I'm a little concerned that implementing the randomNumGenerator in RX is too encouraging of using RX. Considering the number of candidates I've seen try RX now anyway even with encouragement not to, I think this is fine. I also think that Observable.interval is a little obscure so showing users how to run something every second is helpful too. I'm a little worried it's *too* helpful, but if it is, we can always focus more on persistence of the data (which will definitely fill up the hour).

Test Plan:
created a sensor, created a list, filled the list with the random sensor values, subscribed graph to it, everything works dandy